pyproject.toml
[project]
name = "release-copilot-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "agent-framework==1.0.0b251120",
    "azure-identity==1.25.1",
    "openai==2.8.1",
    "python-dotenv==1.2.1",
    "agent-framework-devui>=1.0.0b251120",
    "opentelemetry-api>=1.28.2",
    "opentelemetry-sdk>=1.28.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rc_agent"]

